requirements.txt
Django==4.2.13
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.7.0
gunicorn==22.0.0
Pillow==10.3.0
transformers==4.40.0
torch==2.2.2
sentencepiece==0.2.0
protobuf==4.25.3
accelerate==0.29.3
dj-database-url==2.1.0
